requirements.txt
flask
pandas
numpy
tensorflow
tf-slim
pillow
matplotlib